+++ diesel-wasm-sqlite/src/sqlite_types/to_sql.rs
@@ -4,7 +4,7 @@
 //TODO: CODE CAN BE SHARED (pretty muhch exactly the same)
 use crate::connection::SqliteValue;
 use crate::WasmSqlite;
-use diesel::deserialize::{self, FromSql, Queryable};
+use diesel::deserialize::{self, FromSql};
 use diesel::query_builder::QueryId;
 use diesel::serialize::{self, IsNull, Output, ToSql};
 use diesel::sql_types;
@@ -19,21 +19,39 @@ use diesel::sql_types::SqlType;
 /// raw pointer instead of a reference with a lifetime due to the structure of
 /// `FromSql` because we allocate a string in `read_text` (since SQLite memory is not shared with
 /// us). So this function would
-/// producea  dangling pointer.
-/// `*const` types are not unsafe to _create_ but are unsafe once dereferenced/used
+/// produce a  dangling pointer.
+/*
+// Not posible until we share mem with sqlite. There's no
+// way to avoid an allocation into our host memory until then.
 impl FromSql<sql_types::VarChar, WasmSqlite> for *const str {
     fn from_sql(value: SqliteValue<'_, '_, '_>) -> deserialize::Result<Self> {
+        tracing::debug!("IN FROM SQL");
         let text = value.read_text();
         let text = text.as_str();
         Ok(text as *const _)
     }
 }
+*/
 
+impl FromSql<sql_types::VarChar, WasmSqlite> for String {
+    fn from_sql(value: SqliteValue<'_, '_, '_>) -> deserialize::Result<Self> {
+        Ok(value.read_text())
+    }
+}
+
+/* Not Possible until we share mem with SQLite
 impl Queryable<sql_types::VarChar, WasmSqlite> for *const str {
     type Row = Self;
 
     fn build(row: Self::Row) -> deserialize::Result<Self> {
         Ok(row)
+    }
+}
+*/
+
+impl FromSql<sql_types::Binary, WasmSqlite> for Vec<u8> {
+    fn from_sql(bytes: SqliteValue<'_, '_, '_>) -> deserialize::Result<Self> {
+        Ok(bytes.read_blob())
     }
 }
 
@@ -42,6 +60,7 @@ impl Queryable<sql_types::VarChar, WasmSqlite> for *const str {
 /// impl in terms of `Vec<u8>`, but don't want to allocate. We have to return a
 /// raw pointer instead of a reference with a lifetime due to the structure of
 /// `FromSql`
+/* Not possible until we share mem with SQLite
 impl FromSql<sql_types::Binary, WasmSqlite> for *const [u8] {
     fn from_sql(bytes: SqliteValue<'_, '_, '_>) -> deserialize::Result<Self> {
         let bytes = bytes.read_blob();
@@ -57,6 +76,7 @@ impl Queryable<sql_types::Binary, WasmSqlite> for *const [u8] {
         Ok(row)
     }
 }
+*/
 
 impl FromSql<sql_types::SmallInt, WasmSqlite> for i16 {
     fn from_sql(value: SqliteValue<'_, '_, '_>) -> deserialize::Result<Self> {
